# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Optimal Dump Packing API
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiError {
  error: string;
}

export interface Coordinate {
  x: number;
  y: number;
}

export interface GpsCoordinate {
  lat: number;
  lng: number;
}

export interface TruckProfile {
  id: string;
  name: string;
  width: number;
  length: number;
  turningRadius: number;
  spacingX: number;
  spacingY: number;
  payloadTonnes: number;
}

export interface SpotPoint {
  id: number;
  x: number;
  y: number;
  laneId: number;
  sequenceInLane: number;
  globalSequence: number;
  zoneId: number;
  rotation: number;
  safe: boolean;
}

export interface Lane {
  id: number;
  spotIds: number[];
  approachAngle: number;
}

export interface RotationScore {
  angle: number;
  spotCount: number;
}

export interface DensityMetrics {
  spotCount: number;
  squareGridCount: number;
  improvementPercent: number;
  utilizationEfficiency: number;
  hexPackEfficiency: number;
  squarePackEfficiency: number;
  insetArea: number;
  totalArea: number;
}

export interface PackingPlan {
  spots: SpotPoint[];
  lanes: Lane[];
  polygon: Coordinate[];
  insetPolygon: Coordinate[];
  bestRotation: number;
  rotationScores?: RotationScore[];
  metrics: DensityMetrics;
  truckProfile: TruckProfile;
  generatedAt: string;
}

export interface PackInput {
  /** @minItems 3 */
  polygon: Coordinate[];
  truckProfileId: string;
  customTruck?: TruckProfile;
  ingressAngle?: number;
  rotationStep?: number;
}

export interface GpsPackInput {
  /** @minItems 3 */
  gpsPolygon: GpsCoordinate[];
  truckProfileId: string;
  customTruck?: TruckProfile;
  ingressAngle?: number;
  rotationStep?: number;
}

export interface ExportInput {
  plan: PackingPlan;
  format?: string;
}

export type ExportResultData = { [key: string]: unknown };

export interface ExportResult {
  filename: string;
  data: ExportResultData;
  generatedAt: string;
}

export interface BenchmarkScenario {
  polygonName: string;
  truckId: string;
  hexSpots: number;
  squareSpots: number;
  improvementPercent: number;
  bestRotation: number;
}

export type BenchmarkResultSummary = {
  avgImprovementPercent: number;
  maxImprovementPercent: number;
  testedTrucks: string[];
};

export interface BenchmarkResult {
  scenarios: BenchmarkScenario[];
  summary: BenchmarkResultSummary;
}

export interface KpiItem {
  label: string;
  before: string;
  after: string;
  unit: string;
}

export interface DensityGapResult {
  autonomousSpacing: number;
  staffedSpacing: number;
  densityGapRatio: number;
  rootCause: string;
  innovationDescription: string;
  kpis: KpiItem[];
}

export interface SimulationPolygon {
  id: string;
  name: string;
  polygon: Coordinate[];
  description: string;
}

export interface PresetsResponse {
  truckProfiles: TruckProfile[];
  simulationPolygons: SimulationPolygon[];
}