# lib/api-zod/src/generated/types/benchmarkResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Optimal Dump Packing API
 * OpenAPI spec version: 1.0.0
 */
import type { BenchmarkResultSummary } from './benchmarkResultSummary';
import type { BenchmarkScenario } from './benchmarkScenario';

export interface BenchmarkResult {
  scenarios: BenchmarkScenario[];
  summary: BenchmarkResultSummary;
}