# lib/api-zod/src/generated/types/densityGapResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Optimal Dump Packing API
 * OpenAPI spec version: 1.0.0
 */
import type { KpiItem } from './kpiItem';

export interface DensityGapResult {
  autonomousSpacing: number;
  staffedSpacing: number;
  densityGapRatio: number;
  rootCause: string;
  innovationDescription: string;
  kpis: KpiItem[];
}