# lib/api-zod/src/generated/types/densityMetrics.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Optimal Dump Packing API
 * OpenAPI spec version: 1.0.0
 */

export interface DensityMetrics {
  spotCount: number;
  squareGridCount: number;
  improvementPercent: number;
  utilizationEfficiency: number;
  hexPackEfficiency: number;
  squarePackEfficiency: number;
  insetArea: number;
  totalArea: number;
}