# lib/api-zod/src/generated/types/exportInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Optimal Dump Packing API
 * OpenAPI spec version: 1.0.0
 */
import type { PackingPlan } from './packingPlan';

export interface ExportInput {
  plan: PackingPlan;
  format?: string;
}