# lib/api-zod/src/generated/types/exportResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Optimal Dump Packing API
 * OpenAPI spec version: 1.0.0
 */
import type { ExportResultData } from './exportResultData';

export interface ExportResult {
  filename: string;
  data: ExportResultData;
  generatedAt: string;
}