# lib/api-zod/src/generated/types/gpsPackInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Optimal Dump Packing API
 * OpenAPI spec version: 1.0.0
 */
import type { GpsCoordinate } from './gpsCoordinate';
import type { TruckProfile } from './truckProfile';

export interface GpsPackInput {
  /** @minItems 3 */
  gpsPolygon: GpsCoordinate[];
  truckProfileId: string;
  customTruck?: TruckProfile;
  ingressAngle?: number;
  rotationStep?: number;
}