# lib/api-zod/src/generated/types/kpiItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Optimal Dump Packing API
 * OpenAPI spec version: 1.0.0
 */

export interface KpiItem {
  label: string;
  before: string;
  after: string;
  unit: string;
}