# lib/api-zod/src/generated/types/lane.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Optimal Dump Packing API
 * OpenAPI spec version: 1.0.0
 */

export interface Lane {
  id: number;
  spotIds: number[];
  approachAngle: number;
}