# lib/api-zod/src/generated/types/simulationPolygon.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Optimal Dump Packing API
 * OpenAPI spec version: 1.0.0
 */
import type { Coordinate } from './coordinate';

export interface SimulationPolygon {
  id: string;
  name: string;
  polygon: Coordinate[];
  description: string;
}