# lib/api-zod/src/generated/types/spotPoint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Optimal Dump Packing API
 * OpenAPI spec version: 1.0.0
 */

export interface SpotPoint {
  id: number;
  x: number;
  y: number;
  laneId: number;
  sequenceInLane: number;
  globalSequence: number;
  zoneId: number;
  rotation: number;
  safe: boolean;
}