# lib/api-zod/src/generated/types/truckProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Optimal Dump Packing API
 * OpenAPI spec version: 1.0.0
 */

export interface TruckProfile {
  id: string;
  name: string;
  width: number;
  length: number;
  turningRadius: number;
  spacingX: number;
  spacingY: number;
  payloadTonnes: number;
}